+++ src/main/java/roomescape/payment/infrastructure/toss/client/TossRestClient.java
@@ -1,21 +1,62 @@
 package roomescape.payment.infrastructure.toss.client;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.math.BigDecimal;
+import java.util.Map;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 import roomescape.common.provider.RestClientProvider;
+import roomescape.payment.infrastructure.toss.exception.TossErrorResponse;
+import roomescape.payment.infrastructure.toss.exception.TossInternalException;
+import roomescape.payment.infrastructure.toss.exception.TossPaymentApprovalFailedException;
 
 @Component
 public class TossRestClient {
     private final RestClient restClient;
+    private final ObjectMapper objectMapper;
 
     public TossRestClient(RestClientProvider restClientProvider,
-                          @Value("${toss.baseUrl}") String baseUrl,
-                          @Value("${toss.secretKey}") String secretKey) {
+                          @Value("${external.toss.baseUrl}") String baseUrl,
+                          @Value("${external.toss.secretKey}") String secretKey,
+                          ObjectMapper objectMapper) {
         this.restClient = restClientProvider.createRestClient(baseUrl, secretKey);
+        this.objectMapper = objectMapper;
     }
 
-    public RestClient getRestClient() {
-        return restClient;
+    public void approve(String orderId, BigDecimal amount, String paymentKey) {
+        Map<String, Object> body = Map.of(
+                "orderId", orderId,
+                "amount", amount,
+                "paymentKey", paymentKey
+        );
+
+        restClient.post()
+                .uri("/v1/payments/confirm")
+                .body(body)
+                .retrieve()
+                .onStatus(HttpStatusCode::is4xxClientError,
+                        (request, response) -> {
+                            TossErrorResponse errorResponse = objectMapper.readValue(response.getBody(),
+                                    TossErrorResponse.class);
+                            handleError(errorResponse);
+                        })
+                .onStatus(HttpStatusCode::is5xxServerError,
+                        (request, response) -> {
+                            TossErrorResponse errorResponse = objectMapper.readValue(response.getBody(),
+                                    TossErrorResponse.class);
+                            throw new TossPaymentApprovalFailedException(HttpStatus.INTERNAL_SERVER_ERROR,
+                                    errorResponse.message());
+                        })
+                .toBodilessEntity();
+    }
+
+    private void handleError(TossErrorResponse errorResponse) {
+        if (errorResponse.hasNonUserFacingMessage()) {
+            throw new TossInternalException();
+        }
+        throw new TossPaymentApprovalFailedException(HttpStatus.BAD_REQUEST, errorResponse.message());
     }
 }
