+++ src/main/resources/application.properties
@@ -7,5 +7,7 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.defer-datasource-initialization=true
 spring.jpa.ddl-auto=create-drop
-toss.secretKey=test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
-toss.baseUrl=https://api.tosspayments.com
+external.toss.secretKey=test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
+external.toss.baseUrl=https://api.tosspayments.com
+external.toss.connect-timeout=3000
+external.toss.read-timeout=60000
